refactor and extract import actions
